agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-cartesia
python-dotenv
livekit-plugins-groq
Flask
requests
jsonify
flask-cors
groq